+++ facelog-client-android/src/main/java/net/gdface/facelog/client/IFaceLogClient.java
@@ -127,7 +127,7 @@ public class IFaceLogClient extends IFaceLogDecorator {
 	 * @return
 	 * @see net.gdface.facelog.client.ClientExtendTools#getCmdSnSupplier(net.gdface.facelog.Token)
 	 */
-	public Supplier<Long> getCmdSnSupplier(Token token) {
+	public Supplier<Integer> getCmdSnSupplier(Token token) {
 		return clientTools.getCmdSnSupplier(token);
 	}
 	/**
