+++ facelog-local/src/main/java/net/gdface/facelog/TokenMangement.java
@@ -481,8 +481,8 @@ class TokenMangement implements ServiceConstant {
 	 * 申请一个唯一的命令序列号 
 	 * @param id
 	 */
-	protected long applyCmdSn(int id){
-		long sn = JedisUtils.incr(KEY_CMD_SN);
+	protected int applyCmdSn(int id){
+		int sn = JedisUtils.incr(KEY_CMD_SN);
 		String key = Long.toString(sn);
 		this.cmdSnTable.set(key, id, false);
 		this.cmdSnTable.expire(key);
@@ -491,7 +491,7 @@ class TokenMangement implements ServiceConstant {
 	/** 申请一个唯一的命令响应通道 
 	 * @param id
 	 * @param duration 通道有效时间 大于0有效,否则使用默认的有效期*/
-	protected String applyAckChannel(int id, long duration){
+	protected String applyAckChannel(int id, int duration){
 		String channel = new StringBuffer(ACK_PREFIX)
 				.append(JedisUtils.incr(KEY_ACK_SN))
 				.toString();
@@ -509,8 +509,8 @@ class TokenMangement implements ServiceConstant {
 	 * @param cmdSn
 	 * @return
 	 */
-	protected boolean isValidCmdSn(long cmdSn){
-		return this.cmdSnTable.containsKey(Long.toString(cmdSn));
+	protected boolean isValidCmdSn(int cmdSn){
+		return this.cmdSnTable.containsKey(Integer.toString(cmdSn));
 	}
 	/**
 	 * 判断命令响应通道是否有效
